\name{svmpath-internal}
\title{svm internal functions}
 \alias{Balanced.Initialization}  
 \alias{coef.svmpath}             
 \alias{DowndateKstar}
 \alias{enlist}                   
 \alias{get.svmstep}              
 \alias{modulus}                  
 \alias{InitsvmPath}            
 \alias{OptInit.alpha}            
 \alias{PrintPath}                
 \alias{SnapPath}                 
 \alias{SolveKstar}               
 \alias{StatPath}                 
 \alias{Unbalanced.Initialization}
 \alias{UpdateKstar}
 \description{Internal svmpath functions}
 \author{Trevor Hastie}
 \keyword{internal}
 
